refactor: moved providers to separate folder and split ThemeContext with ThemeProvider
